+++ src/lib/concrete_ono_module.ml
@@ -37,6 +37,18 @@ let clear_screen () : (unit, _) Result.t =
   Buffer.clear buf;
   Ok ()
 
+let config_height () : (Kdo.Concrete.I32.t, _) Result.t =
+  let res = Kdo.Concrete.I32.of_int32 (Game_config.height ()) in
+  Ok (res)
+
+let config_width () : (Kdo.Concrete.I32.t, _) Result.t =
+let res = Kdo.Concrete.I32.of_int32 (Game_config.width ()) in
+Ok (res)
+
+let config_difficulty () : (Kdo.Concrete.I32.t, _) Result.t =
+let res = Kdo.Concrete.I32.of_int32 (Game_config.difficulty ()) in
+Ok (res)
+
 let m =
   let open Kdo.Concrete.Extern_func in
   let open Kdo.Concrete.Extern_func.Syntax in
@@ -49,7 +61,10 @@ let m =
       ("cell_print",   Extern_func (i32  ^->. unit, cell_print));
       ("is_alive",   Extern_func (i32  ^->. i32, is_alive));
       ("newline",      Extern_func (unit ^->. unit, newline));
-      ("clear_screen", Extern_func (unit ^->. unit, clear_screen))
+      ("clear_screen", Extern_func (unit ^->. unit, clear_screen));
+      ("config_height", Extern_func (unit ^->. i32, config_height));
+      ("config_width", Extern_func (unit ^->. i32, config_width));
+      ("config_difficulty", Extern_func (unit ^->. i32, config_difficulty))
     ]
   in
   {