+++ src/tool/cmd_concrete.ml
@@ -11,14 +11,23 @@ let seed =
   let doc = "Set the seed for random number generation." in
   Arg.(value & opt (some int) None & info ["seed"] ~doc)
 
+let configuration =
+  let doc = "Set a config for the game from .sexp file." in
+  Arg.(
+    value & opt (some existing_file_conv) None & info ["config"] ~doc ~docv:"FILE")
+
 let term =
   let open Term.Syntax in
   let+ () = setup_log
   and+ source_file
-  and+ seed in
+  and+ seed
+  and+ configuration in
   (match seed with
    | Some s -> Random.init s
    | None -> Random.self_init ());
+  (match configuration with
+   | Some file -> Game_config.load file;
+   | None -> ());
   Ono.Concrete_driver.run ~source_file |> function
   | Ok () -> Ok ()
   | Error e -> Error (`Msg (Kdo.R.err_to_string e))